802.11: Fixed incorrect destination state after receiving a WLAN ACK.
The 'examples/adhoc/ieee80211' folder contains a simple example: a
wireless host sends ping requests to another. The ping replies always
arrived incorrectly after a zero length backoff period.

The following happened in the MAC module which sends the ping replies:

After the reception of the last WLAN ACK of the last ping reply the MAC
changed to DEFER state. This started an incorrect DIFS period, then a
backoff period, and since there was no frame in the queue, finally the
MAC changed to IDLE state. At the end of this process the backoff counter
was zero. When the next ping reply arrived from the higher layer the MAC
incorrectly continued the remaining backoff period. Thus, the MAC module
always sent out the ping replies after a zero length backoff period.